requirements.txt
google-generativeai
feedparser
beautifulsoup4
requests
jinja2
python-dotenv
python-dateutil
